Fix tests for non-reminder CC cases
