+++ backend/coldchain/src/sensor_state.rs
@@ -9,10 +9,13 @@ use crate::ColdChainError;
 pub struct SensorState {
     pub sensor_id: String,
     pub status: SensorStatus,
-    pub timestamp: NaiveDateTime,
+    #[serde(alias = "timestamp")]
+    pub last_data_localtime: NaiveDateTime,
     pub temperature: Option<f64>,
     #[serde(default)]
-    pub reminder_timestamp: Option<NaiveDateTime>,
+    pub status_start_utc: NaiveDateTime,
+    #[serde(default)]
+    pub last_notification_utc: Option<NaiveDateTime>,
     #[serde(default)]
     pub reminder_number: usize,
 }
@@ -48,16 +51,41 @@ mod test {
     use super::*;
 
     #[test]
-    fn test_parse_status_ok() {
-        let example1 =
-            r#"{ "sensor_id": "1234", "status": "Ok", "timestamp": "2020-01-01T00:00:00" }"#;
+    fn test_parse_status_ok_with_status_start_utc() {
+        let example1 = r#"{ "sensor_id": "1234", "status": "Ok", "last_data_localtime": "2020-01-01T00:00:00", "status_start_utc": "2019-09-01T00:00:00" }"#;
         let result = SensorState::from_string(example1);
+        println!("{:?}", result);
         assert!(result.is_ok());
         let state = result.unwrap();
         assert_eq!(state.sensor_id, "1234");
         assert_eq!(state.status, SensorStatus::Ok);
         assert_eq!(
-            state.timestamp,
+            state.last_data_localtime,
+            NaiveDateTime::parse_from_str("2020-01-01T00:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
+        );
+        assert_eq!(
+            state.status_start_utc,
+            NaiveDateTime::parse_from_str("2019-09-01T00:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
+        );
+
+        // Now save as json
+        let json_string = state.to_json_string().unwrap();
+        let result = SensorState::from_string(&json_string).unwrap();
+        assert_eq!(state, result);
+    }
+
+    #[test]
+    fn test_parse_status_ok() {
+        let example1 =
+            r#"{ "sensor_id": "1234", "status": "Ok", "timestamp": "2020-01-01T00:00:00" }"#;
+        let result = SensorState::from_string(example1);
+        println!("{:?}", result);
+        assert!(result.is_ok());
+        let state = result.unwrap();
+        assert_eq!(state.sensor_id, "1234");
+        assert_eq!(state.status, SensorStatus::Ok);
+        assert_eq!(
+            state.last_data_localtime,
             NaiveDateTime::parse_from_str("2020-01-01T00:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
         );
 
@@ -77,7 +105,7 @@ mod test {
         assert_eq!(state.sensor_id, "1234");
         assert_eq!(state.status, SensorStatus::LowTemp);
         assert_eq!(
-            state.timestamp,
+            state.last_data_localtime,
             NaiveDateTime::parse_from_str("2020-01-01T00:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
         );
 
@@ -97,7 +125,7 @@ mod test {
         assert_eq!(state.sensor_id, "1234");
         assert_eq!(state.status, SensorStatus::HighTemp);
         assert_eq!(
-            state.timestamp,
+            state.last_data_localtime,
             NaiveDateTime::parse_from_str("2020-01-01T00:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
         );
 
@@ -117,7 +145,7 @@ mod test {
         assert_eq!(state.sensor_id, "1234");
         assert_eq!(state.status, SensorStatus::NoData);
         assert_eq!(
-            state.timestamp,
+            state.last_data_localtime,
             NaiveDateTime::parse_from_str("2020-01-01T00:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
         );
 
@@ -129,18 +157,18 @@ mod test {
 
     #[test]
     fn test_parse_status_no_data_reminder() {
-        let example1 = r#"{ "sensor_id": "1234", "status": "NoData", "timestamp": "2020-01-01T00:00:00", "reminder_timestamp": "2020-01-01T01:00:00", "reminder_number":1  }"#;
+        let example1 = r#"{ "sensor_id": "1234", "status": "NoData", "last_data_localtime": "2020-01-01T00:00:00", "last_notification_utc": "2020-01-01T01:00:00", "status_start_utc": "2020-01-01T01:00:00", "reminder_number":1  }"#;
         let result = SensorState::from_string(example1);
         assert!(result.is_ok());
         let state = result.unwrap();
         assert_eq!(state.sensor_id, "1234");
         assert_eq!(state.status, SensorStatus::NoData);
         assert_eq!(
-            state.timestamp,
+            state.last_data_localtime,
             NaiveDateTime::parse_from_str("2020-01-01T00:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
         );
         assert_eq!(
-            state.reminder_timestamp,
+            state.last_notification_utc,
             Some(
                 NaiveDateTime::parse_from_str("2020-01-01T01:00:00", "%Y-%m-%dT%H:%M:%S").unwrap()
             )
