+++ backend/coldchain/src/test/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod process;
